+++ src/app/models/File.ts
@@ -1,11 +1,11 @@
 export class File {
-    Id : string
-    TypeContent: string
-    Resource : string
-    Name : string
-    OriginalName : string
-    Extension : string
-    Size : number
-    DateRegister : Date
-    DateUpdate : Date
+    Id!: string
+    TypeContent!: string
+    Resource!: string
+    Name!: string
+    OriginalName!: string
+    Extension!: string
+    Size!: number
+    DateRegister!: Date
+    DateUpdate!: Date
 }