+++ src/environments/environment.ts
@@ -5,7 +5,18 @@
 
 export const environment = {
   production: false,
-  apiURL:'https://localhost:7286/',
-  apiAuth:'https://localhost:7027/',
-  apiCashier: 'https://localhost:7176/'
+  client_id: 'teste',
+  authorize_url:'http://127.0.0.1:9000/oauth2/authorize?',
+  redirect_uri:'http://127.0.0.1:4200/auth/authorized',
+  scope:'openid profile',
+  response_type:'code',
+  response_mode:'form_post',
+  code_challenge_method:'S256',
+  code_challenge:'T-xl9hy1F5pmy8-bcYIrMNac_TjhwaiVC_o-0Eb0IGo',
+  code_verifier: 'dbBjwS3cYljGnwoSul5pDIKGp9HWgunMXu7zWrbughs',
+  token_url: 'http://127.0.0.1:9000/oauth2/token',
+  grant_type: 'authorization_code'
+  // apiURL:'https://localhost:7286/',
+  // apiAuth:'https://localhost:7027/',
+  // apiCashier: 'https://localhost:7176/'
 };